Erase my root partition on every boot
Using ZFS, make / totally empty on each boot.

Because /home, /boot, and /nix are all on different
datasets, the remaining state of my system is extremely
limited. I can effectively and safely erase / on each
boot without losing much of value.

Previous techniques to do this made / a tmpfs, but I
don't want to store the data in RAM. Writing to a ZFS
dataset, I'm able to "restore" a blank snapshot after
ZFS pools are imported, but before they are mounted.

Right now the only annoying thing I'm noticing is
wireless networks are forgotten on each boot... but
a NixOS option exists for that:

    networking.wireless.networks.<name>