autorandr: dump service, use targets
Autorandr's service depends on some wacky stuff to work.

1. It uses a udev hook to execute the autorandr program

   However, the udev hook triggers too early, or too often depending
   on monitor configurations. This seems to make autorandr trip on
   itself trying to switch. It can also cause autorandr to run only
   once, but before the screens are actually available for autorandr
   to switch to.

2. It uses some Deep Magic to detect "DISPLAY" variables and contact
   the X server for each user. This is pretty weird, too, and depends
   on spelunking through through /proc to find a process with DISPLAY.

Anyway, it seems like a bit of a hack.

This new thing uses systemd user targets instead, which is maybe
sightly less convenient, but more explicit and has not failed on me
yet:

    $ systemctl --user start docked.target

and:

    $ systemctl --user start undocked.target

In the future, I may trigger this target by other local environmental
factors which are _not_ the display, for example a udev event for when
my Thunderbolt dock's specific device ID arrives. Not sure -- that
might expose us to timing problems.

Separately, you don't need to "prime" a resolution if the cable
supports it already.